backend/restaurants/serializers.py
from rest_framework import serializers
from .models import Restaurant

class RestaurantSerializer(serializers.ModelSerializer):
    class Meta:
        model = Restaurant
        fields = ['id', 'name', 'address', 'latitude', 'longitude', 'notes', 'rating', 'created_at']
        read_only_fields = ['id', 'created_at']
